Update types and their directory
