--- server/types/form-body.ts
@@ -1,14 +1,0 @@
-export interface FormBody {
-  formData: {
-    name: string;
-    email: string;
-    message: string;
-  };
-  hCaptcha: {
-    verified: boolean;
-    expired: boolean;
-    token: string;
-    ekey: string;
-    error: Error;
-  };
-}
