+++ utils/types/form-body.ts
@@ -1,0 +1,10 @@
+import { HCaptcha } from '@/utils/types/hcaptcha';
+
+export interface FormBody {
+  formData: {
+    name: string;
+    email: string;
+    message: string;
+  };
+  hCaptcha: HCaptcha;
+}
