+++ utils/types/hcaptcha.ts
@@ -1,0 +1,7 @@
+export interface HCaptcha {
+  verified: boolean;
+  expired: boolean;
+  token: string;
+  eKey: string;
+  error: Error;
+}
